validate and package changes
