+++ cmd/utils.go
@@ -1087,7 +1087,7 @@ func Targz(source, target string) error {
 
 	info, err := os.Stat(source)
 	if err != nil {
-		return nil
+		return err
 	}
 
 	var baseDir string
